fixed styling for mobile view on calculator and laptop
