+++ src/styles/home.module.scss
@@ -1,6 +1,7 @@
 .calculator {
   height: 100vh;
   width: 1900px;
+  margin-bottom: 24px;
   background: url('/images/background.jpg');
   background-repeat: no-repeat;
 }
